kind of clean up how we organize in MathInput
