+++ packages/math-input/src/components/keypad-legacy/index.ts
@@ -1,0 +1,1 @@
+export {default} from "./provided-keypad";
