+++ packages/math-input/src/components/keypad/index.tsx
@@ -7,8 +7,7 @@ import {useEffect} from "react";
 import Key from "../../data/keys";
 import {ClickKeyCallback} from "../../types";
 import {CursorContext} from "../input/cursor-contexts";
-import Tabbar from "../tabbar/tabbar";
-import {TabbarItemType} from "../tabbar/types";
+import Tabbar, {TabbarItemType} from "../tabbar";
 
 import ExtrasPage from "./keypad-pages/extras-page";
 import GeometryPage from "./keypad-pages/geometry-page";
