+++ packages/math-input/src/components/keypad/shared-keys.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import Keys from "../../data/key-configs";
 import {ClickKeyCallback} from "../../types";
 import {CursorContext} from "../input/cursor-contexts";
-import {TabbarItemType} from "../tabbar/types";
+import {TabbarItemType} from "../tabbar";
 
 import {KeypadButton} from "./keypad-button";
