+++ packages/math-input/src/components/prop-types.js
@@ -5,7 +5,6 @@
 import PropTypes from "prop-types";
 
 // NOTE(jared): This is no longer guaranteed to be React element
-// NOTE(matthewc): only seems to be used in Perseus
 export const keypadElementPropType = PropTypes.shape({
     activate: PropTypes.func.isRequired,
     dismiss: PropTypes.func.isRequired,
