+++ packages/math-input/src/components/tabbar/index.ts
@@ -1,0 +1,2 @@
+export {default} from "./tabbar";
+export {TabbarItemType} from "./types";
